+++ src/test/java/PaginaTeste.java
@@ -1,7 +1,9 @@
 
+import testlogin.myFirstTest;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
 import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.support.PageFactory;
 
@@ -16,15 +18,17 @@ public class PaginaTeste {
     {
         System.setProperty("webdriver.gecko.driver", "/ice/Downloads/geckodriver");
         driver = (WebDriver) new FirefoxDriver();
-        driver.get("https://accounts.google.com/signup/v2/webcreateaccount?service=mail&continue=https%3A%2F%2Fmail.google.com%2Fmail%2F&ltmpl=default&flowName=GlifWebSignIn&flowEntry=SignUp");
+        driver.get("https://atlantis.isti.cnr.it:5000/");
     }
 
     @Test
     public void test()
     {
         myFirstTest pag = PageFactory.initElements(driver, myFirstTest.class);
-        pag.preencheDados("Cassioooooo", "Coringao");
+        
         pag.loga();
+        pag.preencheDados("student1@gmail.com", "pass");
+        pag.acessaLoginModal();
         //WebElement ddd;
         //assertNotNull(ddd);
     }
