+++ src/test/java/testlogin/myFirstTest.java
@@ -1,0 +1,48 @@
+package testlogin;
+
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.PageFactory;
+
+public class myFirstTest {
+
+    @FindBy(id = "email")
+    private WebElement email;
+
+    @FindBy(id = "password")
+    private WebElement password;
+
+    @FindBy(id = "signUpButton")
+    private WebElement botao;
+
+    @FindBy(id = "log-in-btn")
+    WebElement botaoLogin;
+
+    private WebDriver driver;
+
+    public myFirstTest(WebDriver driver) {
+        this.driver = driver;
+    }
+
+    public myFirstTest preencheDados(String nome, String sobrenome) {
+        this.email.sendKeys(nome);
+        this.password.sendKeys(sobrenome);
+        PageFactory.initElements(driver, this);
+        return this;
+    }
+
+    public void loga() {
+        driver.findElement(By.linkText("Log in")).click();
+
+        //    botao.click();
+    }
+
+    public void acessaLoginModal() {
+        botaoLogin.click();
+
+    }
+
+}
